+++ crates/affinidi-tdk/common/affinidi-tdk-common/src/lib.rs
@@ -63,8 +63,13 @@ impl TDKSharedState {
         let (secrets_resolver, _) = ThreadedSecretsResolver::new(None).await;
         let client = create_http_client();
         let environment = TDKEnvironment::default();
-        let (authentication, _) =
-            AuthenticationCache::new(1_000, &did_resolver, secrets_resolver.clone(), &client);
+        let (authentication, _) = AuthenticationCache::new(
+            1_000,
+            &did_resolver,
+            secrets_resolver.clone(),
+            &client,
+            None,
+        );
         authentication.start().await;
 
         TDKSharedState {